Make firebase functions use typescript.
