+++ functions/package.json
@@ -1,15 +1,15 @@
 {
   "name": "functions",
   "scripts": {
-    "lint": "./node_modules/.bin/tslint -p tslint.json",
-    "build": "./node_modules/.bin/tsc",
+    "lint": "tslint --project tsconfig.json",
+    "build": "tsc",
     "serve": "npm run build && firebase serve --only functions",
     "shell": "npm run build && firebase experimental:functions:shell",
     "start": "npm run shell",
     "deploy": "firebase deploy --only functions",
     "logs": "firebase functions:log"
   },
-  "main": "src/index.js",
+  "main": "lib/index.js",
   "dependencies": {
     "firebase-admin": "~5.8.1",
     "firebase-functions": "^0.8.1"
